backend/requirements.txt
SQLAlchemy==1.4.50
fastapi==0.104.0
PyMySQL==1.1.0
SQLAlchemy-Utils==0.41.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic_core==2.10.1
pydantic-settings==2.0.3
PyMySQL==1.1.0
mysql-connector-python==8.2.0
python-multipart == 0.0.6
chardet
filelock
packaging
